[DBTablesMonitoring] Remove unused deprecated methods: addHostgroupInfo() and getHostgroupInfoList().
