+++ server/src/DBTablesMonitoring.h
@@ -181,13 +181,6 @@ public:
 	                              const EventsQueryOption &option,
 				      IncidentInfoVect *incidentInfoVect = NULL);
 
-	void addHostgroupInfo(HostgroupInfo *eventInfo)
-	  __attribute__ ((deprecated));
-	void addHostgroupInfoList(const HostgroupInfoList &groupInfoList)
-	  __attribute__ ((deprecated));
-	HatoholError getHostgroupInfoList(HostgroupInfoList &hostgroupInfoList,
-	                      const HostgroupsQueryOption &option)
-	  __attribute__ ((deprecated));
 	HatoholError getHostgroupElementList
 	  (HostgroupElementList &hostgroupElementList,
 	   const HostgroupElementQueryOption &option)
